+++ lib/src/form/components/CustomIdField.tsx
@@ -137,7 +137,6 @@ export function CustomIdField<M extends Record<string, any>, UserType>({
         type: null,
         value: (entity && status === "existing" ? entity.id : entityId) ?? "",
         variant: "filled",
-        disableUnderline: true,
         sx: (theme:Theme) => ({
             minHeight: "64px",
             borderRadius: `${theme.shape.borderRadius}px`